Migrate sphinx-sourceref to Poetry
