+++ tools/sourceref/pyproject.toml
@@ -1,0 +1,16 @@
+[tool.poetry]
+name = "sourceref"
+version = "0.1.0"
+description = "Sphinx extension to conveniently link to source files in a GitHub project"
+authors = ["Philippe Lieser <philippe.lieser@rohde-schwarz.com>"]
+
+[tool.poetry.dependencies]
+python = "^3.10"
+Sphinx = "^6.1.3"
+requests = "^2.25"
+
+[tool.poetry.dev-dependencies]
+
+[build-system]
+requires = ["poetry-core>=1.0.0"]
+build-backend = "poetry.core.masonry.api"
